Extract PIR online-phase benchmark fixture definition into a different header
